@@ -4,22 +4,22 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@emotion/react": "^11.9.3",
-    "@emotion/styled": "^11.9.3",
-    "@mui/material": "^5.9.1",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "bootstrap": "^5.2.0",
-    "i18next": "^20.3.5",
-    "react": "^17.0.2",
-    "react-bootstrap": "^2.5.0",
-    "react-dom": "^17.0.2",
-    "react-i18next": "^11.11.4",
-    "react-icons": "^4.2.0",
-    "react-router-dom": "^6.16.0",
-    "react-scripts": "^3.0.1",
-    "web-vitals": "^1.1.2"
+    "@emotion/react": "^11.14.0",
+    "@emotion/styled": "^11.14.1",
+    "@mui/material": "^7.3.4",
+    "@testing-library/jest-dom": "^6.9.1",
+    "@testing-library/react": "^16.3.0",
+    "@testing-library/user-event": "^14.6.1",
+    "bootstrap": "^5.3.8",
+    "i18next": "^25.6.0",
+    "react": "^19.2.0",
+    "react-bootstrap": "^2.10.10",
+    "react-dom": "^19.2.0",
+    "react-i18next": "^16.2.3",
+    "react-icons": "^5.5.0",
+    "react-router-dom": "^7.9.5",
+    "react-scripts": "^5.0.1",
+    "web-vitals": "^5.1.0"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -48,6 +48,6 @@
     ]
   },
   "devDependencies": {
-    "gh-pages": "^3.2.3"
+    "gh-pages": "^6.3.0"
   }
 }
